### 2021-09-12 #3720 @whedon
[PRE REVIEW]: CategoricalTimeSeries.jl: A toolbox for categorical time-series analysis
**Submitting author:** @johncwok (<a href="http://orcid.org/0000-0001-6266-5575">Corentin  Nelias</a>)
**Repository:** <a href="https://github.com/johncwok/CategoricalTimeSeries.jl" target ="_blank">https://github.com/johncwok/CategoricalTimeSeries.jl</a>
**Version:** 1.0.0
