### 2021-09-12 #3720 @Kevin-Mattheus-Moerman
@johncwok thanks for this submission. I will be looking for a handling editor for this work. In the mean time can you work on fixing those references :point_up:. You can call `@whedon generate pdf` here to update the paper and call `@whedon check references` to check those DOI's. Thanks.
