### 2021-09-12 #3720 @johncwok
Dear Kevin, I checked the existing DOIs and added the missing ones.
