### 2021-09-12 #3720 @jbytecode
Dear @johncwok

You can use the site https://whedon.theoj.org/ to compile your paper and bibtex and then you can let the whedon dump the latest version here.

Do you have any suggestions for potential reviewers?  If yes, please mention them here in this thread (without tagging them with an @ - we don't want to disturb them). In addition, this [list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

In case you are unable to create a suitable list of referees, I will create one for you.

Thank you in advance.
