### 2021-09-12 #3720 @johncwok
@jbytecode I will have a look at potential reviewers this evening, and make test sets. Thanks for the feedback.
I have another question: one of the papers I have in the references (The information bottleneck) doesn't seem to have a DOI. Is it an issue?
