### 2021-09-13 #3720 @jiweiqi
Hi @jbytecode , thanks for your invitation. This package looks cool. However, I am not an expert in analyzing categorical time series (instead I am more familiar with dynamic systems with continuous variable). I would suggest someone else in this field review this paper. Thanks again and I am happy to involve in the future review processes.
