### 2021-09-14 #3720 @bkamins
Hi, I can review it. I have one question and two comments.

The question is: what exactly should I review (code, documentation, the article proof)?

The comments are:
* it would be great to review the package against some release of it. Currently I do not see any releases nor tags for this package.
* the current state of the package when added to an environment does not allow latest version of DataFrames.jl to be installed. It restricts it to 0.19 release. I think it would be good to try to fix that.

Thank you!
