### 2021-09-14 #3720 @jbytecode
> Hi, I can review it. I have one question and two comments.
>
> The question is: what exactly should I review (code, documentation, the article proof)?

@bkamins, Thank you for the reply. When the review starts, there will be 20 checkboxes for each reviewer that indicate the corresponding  review item including the software, the paper, and the documentation. If you accept, I will be here to help you anytime.

>
> The comments are:
>
> * it would be great to review the package against some release of it. Currently I do not see any releases nor tags for this package.
> * the current state of the package when added to an environment does not allow latest version of DataFrames.jl to be installed. It restricts it to 0.19 release. I think it would be good to try to fix that.
>

Since these are about reviewing, you will be able to interact with the author(s) and list a reviewing report during the process. This process will start whenever we set at least two reviewers for the submission.
> Thank you!

Does this answers fit your questions?
