### 2021-09-14 #3720 @johncwok
Dear bkamins,

might I ask what you mean by release? The package was added to the julia registry some 10 months ago (https://github.com/JuliaRegistries/General/tree/master/C/CategoricalTimeSeries).
I am a bit puzzled by the restrictions on the version of dataframes, since I didn't explicitly enforce any. I tried to explicitly make it work for later release, I hope this will fix it.
