### 2021-09-14 #3720 @bkamins
> do these answers fit your questions?

So I wait for the second reviewer. Right?

> might I ask what you mean by release?

In https://github.com/JuliaRegistries/General/blob/master/C/CategoricalTimeSeries/Versions.toml you only have:
```
["0.1.0"]
git-tree-sha1 = "62468caf4b35c78a97dc7a5be9ea068a1f731a27"
```
while in your repo in https://github.com/johncwok/CategoricalTimeSeries.jl/blob/main/Project.toml you currently have:
```
name = "CategoricalTimeSeries"
uuid = "f5116226-a574-5b1a-a4a4-91edc520f498"

authors = ["Corentin Nelias <cwokpoubelle@laposte.net>"]
version = "1.0.0"
```
so as you can see the versions are inconsistent. And there is no tag for 0.1 release on your repository.
Normally it is recommended to make a tag and a release on GitHub for each version of the package.

> I am a bit puzzled by the restrictions on the version of dataframes

The reason is that in IntegerIB.jl in https://github.com/johncwok/IntegerIB.jl/blob/master/Project.toml you have:
```
[compat]
julia = "^1.0"
DataFrames = "^0.19"
```
which disallows latest version of DataFrames.jl.

---

In general I would recommend - as for the reviewed packages to make proper releases also for all packages that the reviewed package depends on as e.g. IntegerIB.jl does not have a proper release either.

To be clear, as is explained here https://github.com/JuliaRegistries/Registrator.jl/#note-on-git-tags-and-github-releases Julia package manager does not depend on GitHub tags, however, this is a good practice to have them and properly manage releases on GitHub (as users will normally check them when using your package). You can have a look at DataFrames.jl as an example how it looks like when this process is followed here: https://github.com/JuliaData/DataFrames.jl/releases.

The reason is that currently when you install your packages Julia installs some very old version of it (almost a year old) and not what is in `main` branch, as it currently uses the `62468caf4b35c78a97dc7a5be9ea068a1f731a27` git-tree-sha1 as a reference what should be installed (and it is dated Oct 28, 2020).
