### 2021-09-14 #3720 @jbytecode
👋👋👋 Dear  @papachristoumarios  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "_CategoricalTimeSeries.jl: A toolbox for categorical time-series analysis_". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/3720).
