### 2021-09-16 #3733 @whedon
[REVIEW]: CategoricalTimeSeries.jl: A toolbox for categorical time-series analysis
**Submitting author:** @johncwok (<a href="http://orcid.org/0000-0001-6266-5575">Corentin  Nelias</a>)
**Repository:** <a href="https://github.com/johncwok/CategoricalTimeSeries.jl" target ="_blank">https://github.com/johncwok/CategoricalTimeSeries.jl</a>
**Version:** v1.1.4
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.16917625" target="_blank">10.6084/m9.figshare.16917625</a>
