### 2021-09-27 #3733 @bkamins
I am not fully sure where I should put the comments, so I put them here:
* I would recommend that all packages (also dependencies) get a proper release
* All packages should have explicit `[compat]` section in their Project.toml (including the `[extras]` part for e.g. CSV.jl)
* The `print_results` does not work under current release of DataFrames.jl (I can help fixing it if needed)
* The examples in the tutorial do not work under current release of CSV.jl (I can help fixing it if needed)
* In general before reviewing the whole documentation and manual should be tested against versions of the packages specified in `[compat]`
* also I would recommend to disambiguate the paths given in the manual (as they are e.g. given in the main package documentation while the files are in the IntegerIB.jl package sources)

Regarding the analytical side of the package:
* the authors mention nominal and ordinal categorical data but do not explain how (or if) the fact that data is ordinal would be used by the package;
* maybe also comment in the documentation how this package could be integrated with CategoricalArrays.jl?

Thank you for working on this.
