### 2021-09-29 #3733 @johncwok
Dear @bkamins,

I addressed your comments, specifically:
- made a release for all involved packages.
- updated the code to be compatible with the most recent versions of ```DataFrames.jl``` and ```CSV.jl```.
- the files linked in the doc are all contained in the ```CategoricalTimeSeries.jl``` repo, and not pointing to other repos anymore.
- I tested the functions with the most recent versions of every packages (as specified in ```compat```), and they are working.
- I adapted the description in the doc. The reason why I wrote about 'nominal' and 'ordinal' was just as a sort of introduction regarding what type of data are referred to as 'categorical'.
In itself the package is agnostic about the precise type of data since none of the methods are relying on a pre-existing ordering. The only requirement is that the data should be discrete. One would still get results with continuous data, but these would most probably be nonsensical.

I have one question\need help:
- The reason why I didn't have any ```compat``` entry is because when I do add them, then the package doesn't pass the test on the travis virtual machine, even though it works perfectly otherwise. I end up having error like
```Julia
Unsatisfiable requirements detected for package foo [xxxxx]:

 foo [xxxxx] log:

 ├─possible versions are: [0.1.0-0.1.2, 0.2.0-0.2.4, 0.3.0, 1.0.0-1.0.1, 1.1.0-1.1.1, 1.2.0-1.2.4, 1.3.0-1.3.2, 1.4.0-1.4.5] or uninstalled
 ├─restricted to versions 1.4.5-1 by an explicit requirement, leaving only versions 1.4.5
 └─restricted by julia compatibility requirements to versions: [0.1.0-0.1.2, 0.2.0-0.2.4, 0.3.0, 1.0.0-1.0.1, 1.1.0-1.1.1] or uninstalled — no versions left
```

When I try to fix it, another package will throw this error. For example, if I fix it for ```FFTW```, then ```Plots``` errors, and if I fix it for ```Plots``` then another one bugs and it is a never ending cycle. I search for a solution but  couldn't fix this so I just removed all "problematic" compats in order to have a working package.
Do you know what causes this and how I can solve it? There are already similar issues on the discourse but they are not resolved or do not solve my issue.
