### 2021-09-29 #3733 @bkamins
> then the package doesn't pass the test on the travis virtual machine, even though it works perfectly otherwise

The reason is the following. You seem not to have done the package releases properly. For instance IntegerIB.jl package has 0.1 tag only made yesterday. However, in Julia Global Registry, here https://github.com/JuliaRegistries/General/blob/master/I/IntegerIB/Versions.toml you see that 0.1 tag was added 13 months ago with git-tree-sha1 `786b9878a5889e1feec38510e694016c3bb76c47`.

This makes Julia Package manager install the IntegerIB.jl package from 13 months ago not from yesterday when you ask for it being installed (you can check it on a CLEAN environment - without any packages installed in a federated package repository - this is exactly what happens on Travis).

Note that Julia does not look at what is the tag in the GitHub repository for the release but it looks at  git-tree-sha1 in https://github.com/JuliaRegistries/General to resolve package versions. You can read how new releases of packages should be registered in https://github.com/JuliaRegistries/Registrator.jl
