### 2021-09-29 #3733 @johncwok
So re-registering all the packages to their current version should solve this right?
