### 2021-09-29 #3733 @bkamins
No - you have to bump the version number and then register. The problem is exactly the fact that you use the same version number for different states of package GitHub repository (Julia will not allow registering the same version number twice as it would be ambiguous).

Also note that this has to be done sequentially. If A depends on B and B on C then you have to:
* bump version of C
* register new version of C
* change B to depend on new version of C
* bump version of B
* register new version of B
* change A to depend on new version of B
* bump version of A
* register new version of A
