### 2021-09-30 #3733 @johncwok
@bkamins I took care of it, and it is passing CI now. I don't know how long it will take until the re-registration is done though.
