### 2021-09-30 #3733 @bkamins
> I don't know how long it will take until the re-registration is complete though.

I have checked your PRs to General Registry. As you can see your packages (for different reasons) fail the registration process.
If you have any questions regarding the registration please ask me and I will try to help.

Otherwise - could you please let me know when the registration is finished - thank you!
(in particular at least one of the packages is missing license and having proper licensing is also a JOSS requirement)
