### 2021-09-30 #3733 @johncwok
Do you know why the auto-merge from ```MotifRecognition.jl``` failed? The license is there, and the error message is totally incomprehensible to me.
I have added the missing license to ```SpectralEnvelope.jl```.
