### 2021-09-30 #3733 @bkamins
> Do you know why the auto-merge from MotifRecognition.jl failed?

As you can see in https://github.com/JuliaRegistries/General/pull/45831#issuecomment-931248106 there are two problems:
* your `[compat]` section is incomplete (missing entry for DataStructures.jl)
* your version numbering is skipping versions. Since currently known to Julia Registry version of this package is 0.1 then it expected that the patch release would have number 0.1.1, while you have made the release with 0.1.11 version, which skips many versions in-between (like 0.1.1, 0.1.2, 0.1.3, ..., 0.1.10)
