### 2021-09-30 #3733 @johncwok
I have added the ```[compat]``` for DataStructures.jl.

I guess I will just let you know when the re-registration is complete then.
