### 2021-10-06 #3733 @johncwok
Thanks! I actually already fixed the ```compat```, but as bkamins stated earlier, this was not the only reason why the automatic registration was rejected by the bot. It is because my version numbering was not done according to the convention.
Do you know how to trigger the registration bot again? I don't want to open another pull request for that and create extra work for the people dealing with the registrations by hand.
