### 2021-10-06 #3733 @jbytecode
how about using

`@ JuliaRegistrator register`

in the registration issue as stated in https://github.com/JuliaRegistries/Registrator.jl or using the JuliaHub?
