### 2021-10-06 #3733 @bkamins
Yes - re-triggering JuliaRegistrator should work.
