### 2021-10-06 #3733 @johncwok
I get the following message from the registrator:
Error while trying to register: "Tag with name v1.0.1 already exists and points to a different commit
I made increments of the version when I fixed various bugs but did not register them, and now I can't auto merge because I cannot use an earlier version number that would meet the guidelines.
