### 2021-10-06 #3733 @bkamins
Have you tried removing the tag? Also which exact repository you are referring to? (and could you please share the link to the commit at which you have triggered the registrator). Thank you!
