### 2021-10-06 #3733 @johncwok
This is the link: https://github.com/johncwok/CategoricalTimeSeries.jl/commit/dbf270a8acfd43fa32cf531795d8358a0bc47cc4
I'll try to remove the tag (I am a bit unsure how) and see if it works.
