### 2021-10-07 #3733 @bkamins
> external files in examples

These files are shipped along with the package installation, so I think it is OK to use them.
