### 2021-10-07 #3733 @bkamins
> Do you mean that this should be stated in the documentation

It should be stated in the documentation. If you do `using SerialDependence` you have to install SerialDependence.jl earlier. So either do not do `using SerialDependence` or instruct the user that it should be installed. I would prefer that `using CategoricalTimeSeries` should be enough and it should re-export all functionalities provided by dependencies, but it is of course up to you how you design it. The crucial thing is that it should be possible to just copy-paste the code from the documentation and it should work.
