### 2021-10-07 #3733 @bkamins
you have `dirname` twice in `joinpath(dirname(dirname(pathof(CategoricalTimeSeries))), "test", "pewee.txt")`
