### 2021-10-07 #3733 @johncwok
>
>
> Your bounds for versions of Plots.jl in CategoricalTimeSeries.jl are incorrect. They disallow installing latest Plots.jl and also disallow installing latest SerialDepencence.jl version (which is required for things to work AFAICT).
>
> Can you please share Project.toml and Manifest.toml that you are using under which all works, as AFAICT, it is impossible to install the version configuration of the packages that you say that you have installed.

I was using the old version since I thought that the new one hadn't been merged yet. I changed the Plots.jl compat entry,  now the package can be installed from the url without error.
