### 2021-10-07 #3733 @bkamins
The easiest thing to do is to change `algorithm` into a keyword argument. Then all will be clean. Such a change is breaking so you would need to go for 0.2 release then.
