### 2021-10-09 #3733 @jbytecode
I reinstalled the package from the repository using

```julia
] add CategoricalTimesSeries
```

and at least one test doesn't pass:

```julia
     Testing Running tests...
Test Failed at /Users/istar1/.julia/packages/CategoricalTimeSeries/lypxN/test/runtests.jl:17
  Expression: round(((spectral_envelope(DNA))[2])[5]; digits = 3) == round(0.175; digits = 3)
   Evaluated: 0.003 == 0.175
ERROR: LoadError: There was an error during testing
in expression starting at /Users/istar1/.julia/packages/CategoricalTimeSeries/lypxN/test/runtests.jl:17
ERROR: Package CategoricalTimeSeries errored during testing
```
