### 2021-10-09 #3733 @jbytecode
Addition: Your tests are somehow verbose as it prints

```julia
merged clusters 1 and 2
merged clusters 1 and 3
```

and I think it comes from `IntegerIB.jl `

```julia
if calc_metrics(tmp)[4] < calc_metrics(best)[4]
                    best = deepcopy(tmp)
                    mergedt1 = t1
                    mergedt2 = t2
                    anybetter = true
                    print("merged clusters $mergedt1 and $mergedt2 \n")
                end
```

It seems it is for debugging, but I think throwing errors or using `@info`, `@error`, `@warn` is more convenient if you really want inform the user. Since it is not really essential at this time and maybe you would be willing to change it in forward releases.
