### 2021-10-10 #3733 @johncwok
I have updated the documentation, the docstring for ```bootstrap_CI``` as well as the ```plot_motif``` example.  I also ran again every example to make sure the output and the example show the same thing.
