### 2021-10-10 #3733 @bkamins
@jbytecode - I have finished the review in the first comment above (I guess this is what is required).

@johncwok some issues I would recommend to resolve in the future (they are non-blocking):
* still you use specification of types as `Array{any, 1}` (incorrect), `Array{Float, 1}` (incorrect), or `Array{Any, 1}` (correct but for sure not what you meant) in the documentation;
* the testset for the package is not comprehensive enough IMO;
* the change in the plot for `bootstrap_CI`: it is not clear for me if the old plot was incorrect? (I see you have changed the plot in the docs, but I do not fully understand what has changed in the code so that the plot changed)

Best,
Bogumił
