### 2021-10-27 johncwok/CategoricalTimeSeries.jl#7 @felixcremer
get_mapping with too high or too low frequency values throws BoundsError
When I try to compute the get_mapping function with values for freq at the edges I get the following error:

```julia
julia> mappings = get_mappings(data, f[21])
ERROR: BoundsError: attempt to access 600-element Vector{Float64} at index [-3:45]
Stacktrace:
 [1] throw_boundserror(A::Vector{Float64}, I::Tuple{UnitRange{Int64}})
   @ Base ./abstractarray.jl:651
 [2] checkbounds
   @ ./abstractarray.jl:616 [inlined]
 [3] getindex
   @ ./array.jl:807 [inlined]
 [4] get_mappings(data::Matrix{Any}, freq::Float64; m::Int64)
   @ CategoricalTimeSeries ~/Documents/Programmieren/CategoricalTimeseriesreview/dev/CategoricalTimeSeries/src/SpectralEnvelope.jl:232
 [5] get_mappings(data::Matrix{Any}, freq::Float64)
   @ CategoricalTimeSeries ~/Documents/Programmieren/CategoricalTimeseriesreview/dev/CategoricalTimeSeries/src/SpectralEnvelope.jl:227
 [6] top-level scope
   @ REPL[33]:1
```

If you would like to restrict it to the boundaries of the `se` vector you can use max(peak_pos-window,begin):min(peak_pos+window,end) in the indexing.
This would lead to different search window sizes at the edges and I am not sure, whether this would be a problem.

Part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/3733
