### 2021-10-27 johncwok/CategoricalTimeSeries.jl#8 @felixcremer
get_mapping for maxima of the spectral envelope gives not the eigvecs of this position
From the documentation I understand, that the get_mapping function searches for the local maxima near the frequency I specified and gives the same results I would also get in the eigvec vector from the `spectral_envelope` call, but as a dictionary.
Therefore I assumed, that I would get the same results when I start with the global maximum of the spectral envelope, but for the example data I get the following.
Maybe this is just a misunderstanding on my part.
I would have most of all expected, that the reported strength of peak from the get_mappings function is the same as the maxse value.

```julia
julia> maxse, maxpos = findmax(se)
(0.02204087285907974, 401)

julia> get_mappings(data,f[maxpos])
(true_pos, peak_pos) = (399, 401)
position of peak: 0.33 strengh of peak: 0.01
Dict{SubString{String}, Float64} with 4 entries:
  "A" => 0.54
  "T" => -0.57
  "C" => 0.0
  "G" => 0.62

julia> eigvecs[:,maxpos]
3-element Vector{Float64}:
 -0.586861830502818
  0.5964570350305273
  0.5475693538370039
```

Part of the review for JOSS openjournals/joss-reviews#3733
