### 2021-11-01 johncwok/CategoricalTimeSeries.jl#9 @felixcremer
Document that you could get different local maxima with IB_optimize!
When I rerun the example in the data clustering documentation, this is not always giving the same result.
This is not stated in the documentation of IB, and was surprising for me.
This is mentioned in the search_optima docstring, but it would be better to also mention this in the IB docstring.
Especially the Bach chorale example confused me, because you state the results as given, but I got different results.
Maybe you could use search_optima in the Bach example, before you start interpreting the results.

As a side note:
I would put the examples further on top of the documentation page, either on top of the further usage or at least on top of the additional functions sections, because  examples are  the first thing a user searches for.

Part of the review for JOSS openjournals/joss-reviews#3733
