### 2021-11-01 johncwok/CategoricalTimeSeries.jl#10 @felixcremer
Link to the paper in motif recognition docs
It would be good to link to the JEREMY BUHLER and MARTIN TOMPA's paper "Finding Motifs Using Random Projections" paper that is mentioned in the motif docs.

Part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/3733
