### 2021-11-01 #3733 @felixcremer
@johncwok  Thank you for this nice package.

@jbytecode  - I have finished the review in the first comment above.
It would be good to address the outstanding issues, that  I opened on the repo and linked to this issue.

I think you could convert the list of open improvements in the Contributing.md files into github issue, to make it easier for others to comment and also work on them.

You could set up coveralls.io or codecov.io to make it easier to see the state of testing on the package.
