### 2021-11-01 #3733 @johncwok
@felixcremer Thank you for the thorough code testing and the suggestions for doc improvement!
I have never worked with coveralls.io or codecov.io so far, but I will look into it.
