### 2021-11-01 johncwok/CategoricalTimeSeries.jl#10 @johncwok
Done! I also added a link to the Spectral envelope paper as well as the categorical time-series analysis from Christian Weiss.
