### 2021-11-01 johncwok/CategoricalTimeSeries.jl#9 @johncwok
I added ```search_optima``` in the example for more consistent result and adapted the text.
