### 2021-11-01 #3733 @jbytecode
Dear @johncwok

I sent a pull request that includes possible correction in the `paper.md` file:

https://github.com/johncwok/CategoricalTimeSeries.jl/pull/12

- Please check the corrections and if you are agree with the correction, please apply them.
