### 2021-11-01 #3733 @johncwok
Dear @jbytecode

I am fine with all corrections. I see you changed "time-series" to "time series", is this the correct way of writing it? I have seen "time-series", "timeseries" and "time series". After checking, all of these seem to be valid so I am always a bit unsure how to write it.
