### 2021-11-01 #3733 @jbytecode
@johncwok you mostly use `time-series`, so I changed all other use of this term (time series, timeseries) to `time-series`.
