### 2021-11-02 #3733 @johncwok
Dear @jbytecode

I have done a final check of the paper and made two minor word adjustments, I am finished now.

The latest release (v1.1.4) is archived on figshare with the following DOI [10.6084/m9.figshare.16917625](10.6084/m9.figshare.16917625)
