### 2021-11-02 #3733 @danielskatz
@johncwok - can you update the title of the figshare archive to match the title of the paper: "CategoricalTimeSeries.jl: A toolbox for categorical time-series analysis"?
