### 2021-11-02 #3733 @danielskatz
I've also suggested some small changes in the paper in https://github.com/johncwok/CategoricalTimeSeries.jl/pull/13 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publishing.
