### 2021-11-02 #3733 @johncwok
Dear @danielskatz,

I have no objections to the small changes you made to the paper, thanks for the suggestions.
I changed the title of the figshare archive to match the paper.
