### 2021-11-02 #3733 @danielskatz
At least for me right now, the DOI is not resolving - sometimes it takes longer than other time.

Once it does, I will close this issue.
